% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss2dist.R
\name{cleandist}
\alias{cleandist}
\title{Pranger}
\usage{
cleandist(x)
}
\arguments{
\item{x}{[\code{matrix}] A dissimilarity matrix}
}
\value{
[\code{dist}] A distance matrix from class \code{dist}
}
\description{
Clean the dissimilarity matrix in input by converting it into a distance
matrix
}
