% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr_resampling.R
\name{hdr_resampling}
\alias{hdr_resampling}
\title{Bases on highest density regions to re-sample on sparse region}
\usage{
hdr_resampling(
  data,
  probs = c(95, 75, 50),
  min_region_size,
  nb_bootst = ceiling(sqrt(length(data))),
  aggregation,
  den = NULL,
  h = hdrbw(BoxCox(data, lambda), mean(probs)),
  lambda = 1,
  nn = 5000,
  all.modes = FALSE
)
}
\arguments{
\item{data}{[vector(1)] Numeric vector containing data. If x is missing then
den must be provided, and the HDR is computed from
the given density.}

\item{probs}{[vector(1)] Probability coverage required for HDRs}

\item{min_region_size}{[integer(1)] Minimal region size.}

\item{nb_bootst}{[integer] Number of repetitions required to aggregate the
bootstrap samples. Set to ceiling(sqrt(n)) if not
provided, with n the number of observations}

\item{aggregation}{[function] Aggregation function in the case of "boostaggr".
Default is the mean function. You can pass your
on aggregation function}

\item{den}{[density(1)] Density of data as list with components x and y. See
package hdrcde for more details.}

\item{h}{[numeric(1)] Optional bandwidth for calculation of density.}

\item{lambda}{[numeric(1)] Box-Cox transformation parameter where
0 <= lambda <= 1}

\item{nn}{[integer(0)] Number of random numbers used in computing f-alpha
quantiles.}

\item{all.modes}{[boolean(1)] Return all local modes or just the global mode?}
}
\value{
[vector(1)] a vector with both original and synthetic data
}
\description{
Bases on highest density regions to re-sample on sparse region
}
\examples{
set.seed(123)
data <- c(rf(1000, 10, 90),
         rnorm(1000, 4, 0.5),
         rnorm(1000, 7, 0.5),
         rnorm(1000, 10, 0.5))
probs <- c(95, 75, 50)
set.seed(123)
hdr_rspl <- hdr_resampling (data = data,
                           probs = probs,
                           min_region_size = 100,
                           nb_bootst = ceiling(sqrt(length(data))),
                           aggregation = median)
}
\author{
Cesaire J. K. Fouodo
}
