% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_regions.R
\name{higher_regions}
\alias{higher_regions}
\title{Compute higher density regions for a given vector confidences}
\usage{
higher_regions(
  data,
  probs = c(95, 75, 50),
  den = NULL,
  h = hdrbw(BoxCox(data, lambda), mean(probs)),
  lambda = 1,
  nn = 5000,
  all.modes = FALSE
)
}
\arguments{
\item{data}{[vector(1)] Numeric vector containing data. If x is missing then
den must be provided, and the HDR is computed from
the given density.}

\item{probs}{[vector(1)] Probability coverage required for HDRs}

\item{den}{[density(1)] Density of data as list with components x and y. See
package hdrcde for more details.}

\item{h}{[numeric(1)] Optional bandwidth for calculation of density.}

\item{lambda}{[numeric(1)] Box-Cox transformation parameter where
0 <= lambda <= 1}

\item{nn}{[integer(0)] Number of random numbers used in computing f-alpha
quantiles.}

\item{all.modes}{[boolean(1)] Return all local modes or just the global mode?}
}
\value{
[list] List of highest density regions found with correspponding sum-
               mary description
}
\description{
Compute higher density regions for a given vector confidences
}
\examples{
set.seed(123)
data <- c(rf(500, 10, 90), rnorm(500, 4, 1))
probs <- c(50, 75, 95)
hr_res_list <- higher_regions(data = data, probs = probs)
}
