% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_regions.R
\name{highest_regions}
\alias{highest_regions}
\title{Compute higher density regions for a given vector confidences}
\usage{
highest_regions(
  data,
  probs = c(95, 75, 50),
  min_region_size = NULL,
  den = NULL,
  h = hdrbw(BoxCox(data, lambda), mean(probs)),
  lambda = 1,
  nn = 5000,
  all.modes = FALSE
)
}
\arguments{
\item{data}{[vector(1)] Numeric vector containing data. If x is missing then
den must be provided, and the HDR is computed from
the given density.}

\item{probs}{[vector(1)] Probability coverage required for HDRs}

\item{min_region_size}{[integer(1)] Minimal region size.}

\item{den}{[density(1)] Density of data as list with components x and y. See
package hdrcde for more details.}

\item{h}{[numeric(1)] Optional bandwidth for calculation of density.}

\item{lambda}{[numeric(1)] Box-Cox transformation parameter where
0 <= lambda <= 1}

\item{nn}{[integer(0)] Number of random numbers used in computing f-alpha
quantiles.}

\item{all.modes}{[boolean(1)] Return all local modes or just the global mode?}
}
\value{
The highest density region
}
\description{
Compute higher density regions for a given vector confidences
}
\examples{
set.seed(123)
data <- c(rf(500, 10, 90), rnorm(500, 4, 1))
probs <- c(50, 75, 95)
highestr_res <- highest_regions(data = data,
                                min_region_size = 100,
                                probs = probs)
}
