% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predicted_forest_distance.R
\name{predicted_forest_distance}
\alias{predicted_forest_distance}
\title{Pranger}
\usage{
predicted_forest_distance(predictions, forest)
}
\arguments{
\item{predictions}{[\code{ranger}] An ranger predicted object of terminal nodes}

\item{forest}{[\code{ranger}] A ranger forest}
}
\value{
[\code{matrix}] Matrix of dissimilarities over the forest
}
\description{
Computes dissimilarity matrix over the forest. Note
that this function will parallelize the pairwise computations of the
distances between the nodes on the available CPUs.
}
\examples{
\dontrun{
set.seed(1234)
train.idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris.train <- iris[train.idx, ]
iris.test <- iris[-train.idx, ]
rg.iris <- ranger(Species ~ ., data = iris.train)
pred.iris <- predict(rg.iris, data = iris.test, type = "terminalNodes")
 iris_dist <- predicted_forest_distance(
   forest = rg.iris,
   predictions = pred.iris)
}
}
\author{
Cesaire J. K. Fouodo
}
