% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_predicted_forest_one_tree.R
\name{predicted_forest_distance_parall}
\alias{predicted_forest_distance_parall}
\title{Pranger}
\usage{
predicted_forest_distance_parall(index_tree, predictions, forest)
}
\arguments{
\item{index_tree}{[\code{integer}] Index of concerned tree}

\item{predictions}{[\code{ranger}] An ranger predicted object of terminal
nodes}

\item{forest}{[\code{ranger}] A ranger forest}
}
\value{
[\code{matrix}] Matrix of dissimilarities for the concerned tree
}
\description{
Computes dissimilarities for one tree of the forest
}
\examples{
\dontrun{
library(ranger)
set.seed(1234)
train.idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris.train <- iris[train.idx, ]
iris.test <- iris[-train.idx, ]
rg.iris <- ranger(Species ~ ., data = iris.train)
pred.iris <- predict(rg.iris, data = iris.test, type = "terminalNodes")
 iris_dist <- predicted_forest_distance_parall(
   forest = rg.iris,
   predictions = pred.iris)
   }
}
\author{
Cesaire J. K. Fouodo
}
