% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resampling}
\alias{resampling}
\title{resampling builts a two-classes classification problem using a resampling
           strategy. This implementation is comparable to
           that of Shi and Hovarth (2006)}
\usage{
resampling(data, strategy)
}
\arguments{
\item{data}{[data.frame(1)] A \code{data.frame} of original dataset}

\item{strategy}{[character(1)] Name of strategy to be used. Most be element of
"boostrepl", "boostwithoutrepl", "boostbayes",
"unif", "normal" or "binomial"}
}
\value{
[data.frame(1)] A \code{data.frame} of a two-classes classification
                        problem
}
\description{
resampling builts a two-classes classification problem using a resampling
           strategy. This implementation is comparable to
           that of Shi and Hovarth (2006)
}
\examples{
    set.seed(1234)
   data_synt <- resampling(
                    data = data.frame(x = rnorm(10),
                                      y = rnorm(10)),
                    strategy = "boostrepl"
                  )
}
\author{
Cesaire J. K. Fouodo
}
