% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_search_parents.R
\name{search_parents}
\alias{search_parents}
\title{Pranger}
\usage{
search_parents(ranger_tree)
}
\arguments{
\item{ranger_tree}{[\code{ranger.tree}] A ranger tree\cr}
}
\value{
[\code{data.table}]
  A \code{data.table} containing parent nodes of each terminal node
}
\description{
Search parent nodes of each terminal node from a tree
}
\examples{
child_nodes <- list(
                    c(1, 0, 3, 5, 0, 0, 7, 0, 0),
                    c(2, 0, 4, 6, 0, 0, 8, 0, 0)
                  )
 parent_nodes <- search_parents(ranger_tree = child_nodes)
}
\author{
Cesaire J. K. Fouodo
}
