% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_shi_ranger_forest.R
\name{shi_ranger_forest}
\alias{shi_ranger_forest}
\title{Pranger}
\usage{
shi_ranger_forest(predictions)
}
\arguments{
\item{predictions}{[\code{predict.ranger}] Terminal nodes predicted by
\code{predict.ranger}}
}
\value{
[\code{matrix}] Dissimilarity matrix according to Shi and Hovarth (2006)
}
\description{
Computes dissimilarities between  individuals based on the approach of
Shi and Hovarth (2006) over a forest. For each tree, dissimilarity between
two individuals is increased to one if they don't belong to the same
terminal nodes. Dissimilarities over the forest is obtained by summing
dissimilarities obtained from each tree.
}
\examples{
library(ranger)
set.seed(1234)
train.idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris.train <- iris[train.idx, ]
iris.test <- iris[-train.idx, ]
rg.iris <- ranger(Species ~ ., data = iris.train, num.trees = 5)
pred.iris <- predict(rg.iris, data = iris.test, type = "terminalNodes")
shi_dist <- shi_ranger_forest(predictions = pred.iris)
}
