% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_shi_ranger_one_tree.R
\name{shi_ranger_one_tree}
\alias{shi_ranger_one_tree}
\title{Pranger}
\usage{
shi_ranger_one_tree(tree_index = 1, predictions, init_dist = 0)
}
\arguments{
\item{tree_index}{[\code{integer}] Tree index for which the dissimilarity matrix is required}

\item{predictions}{[\code{predict.ranger}] Terminal nodes predicted by
\code{predict.ranger}}

\item{init_dist}{[\code{integer}] Initial distance between in-of-bag
individual. Required if oob = TRUE.}
}
\value{
[\code{matrix}] Dissimilarity matrix according to Shi and Hovarth (2006)
}
\description{
Given a tree, this function computes dissimilarities between  individuals
based on the approach of Shi and Hovarth (2006). That is, dissimilarity
between two individuals is increased to one if they don't belong to the same
terminal nodes
}
\examples{
library(ranger)
set.seed(1234)
train.idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris.train <- iris[train.idx, ]
iris.test <- iris[-train.idx, ]
rg.iris <- ranger(Species ~ ., data = iris.train)
pred.iris <- predict(rg.iris, data = iris.test, type = "terminalNodes")
shi_dist <- shi_ranger_one_tree(tree_index = 1,
                 predictions = pred.iris)
}
