% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_tree_distance.R
\name{tree_distance}
\alias{tree_distance}
\title{Compute distances between all terminal nodes of a given tree}
\usage{
tree_distance(ranger_tree = child_nodes)
}
\arguments{
\item{ranger_tree}{[ranger.tree] A ranger tree}
}
\value{
[matrix(1)] A matrix of dissimilarities
}
\description{
Compute distances between all terminal nodes of a given tree
}
\examples{
child_nodes <- list(
                    c(1, 0, 3, 5, 0, 0, 7, 0, 0),
                    c(2, 0, 4, 6, 0, 0, 8, 0, 0)
                  )
parent_nodes <- search_parents(ranger_tree = child_nodes)
nd <- nodes_distance(
parent_nodes = parent_nodes,
node1 = 5, node2 = 1)
tree_distance(
ranger_tree = child_nodes
)
}
\author{
Cesaire J. K. Fouodo
}
