% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_urf_test.R
\name{urf_test}
\alias{urf_test}
\title{urf_test computes UNAIR variable importances and uses them for the
Janitza et al. (2018) high-dimensional testing procedure.}
\usage{
urf_test(data, target = "target", resampling_seed, ...)
}
\arguments{
\item{data}{[\code{data.frame}] original data}

\item{target}{[\code{character}] name of the artificial target variable}

\item{resampling_seed}{[\code{integer}] resampling seed}

\item{...}{more parameters to pass to ranger}
}
\value{
[\code{data.table}] with variable importances and p-values
}
\description{
urf_test computes UNAIR variable importances and uses them for the
Janitza et al. (2018) high-dimensional testing procedure.
}
\references{
Janitza, S, Celik, E, Boulesteix, AL. (2018). A computationally fast variable importance test for random forests for high-dimensional data. Adv Data Anal Classif.; doi.org: 10.1007/s11634-016-0276-4
Cesaire J. K. Fouodo, Inke R. König Silke Szymczak (2022) Computing variable importance with unsupervised random forests. In review process.
}
\author{
Cesaire J. K. Fouodo
}
