% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_nodes_distance.R
\name{nodes_distance}
\alias{nodes_distance}
\title{Compute distance between two terminal nodes}
\usage{
nodes_distance(parent_nodes, node1, node2)
}
\arguments{
\item{parent_nodes}{[data.table(1)] \code{data.table} of parent nodes}

\item{node1}{[integer(1)] A node}

\item{node2}{[integer(1)] A node}
}
\value{
[integer(1)] distance between node1 and node2
}
\description{
Compute distance between two terminal nodes
}
\examples{
child_nodes <- list(
                    c(1, 0, 3, 5, 0, 0, 7, 0, 0),
                    c(2, 0, 4, 6, 0, 0, 8, 0, 0)
                  )
parent_nodes <- search_parents(ranger_tree = child_nodes)
nd <- nodes_distance(
parent_nodes = parent_nodes,
node1 = 5, node2 = 1)
}
\author{
Cesaire J. K. Fouodo
}
