% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pranger.R
\name{pranger}
\alias{pranger}
\title{Pranger}
\usage{
pranger(
  data,
  strategy,
  nb_bootst = NULL,
  approach = "deep",
  aggregation = mean,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{[\code{data.frame}] A \code{data.frame} of original dataset}

\item{strategy}{[\code{character}] Name of the resampling strategy to be used.
    Most be element of
"boostrepl", "boostwithoutrepl", "boostbayes",
"unif", "normal", "binomial" or "boostaggr"}

\item{nb_bootst}{[\code{integer}] Number of repetitions required to aggregate the
bootstrap samples. Set to ceiling(sqrt(n)) if not
provided, with n the number of observations}

\item{approach}{[\code{character}] One of "shi" or "deep". If "shi", then
the approach of Shi and  Hovarth (2006) is called to compute dissimilarities.
Else, if "deep" the approach of Fouodo et al. (2021) is called.}

\item{aggregation}{[\code{function}] Aggregation function in the case of
"boostaggr". Default is the mean function.
You can pass your on aggregation function}

\item{verbose}{[\code{boolean}] If TRUE, verbose}

\item{...}{further parameters to be passed to \code{ranger}}
}
\value{
[\code{matrix}] Matrix of dissimilarities.
                        Note: You can use the function \code{cleandist} to
                        convert the dissimilarity matrix into a distance
                        matrix
}
\description{
Pranger bases on unsupervised random forests generated with \code{ranger}
to compute proximities between individuals.Two approaches are supported. The
first approach of Shi and Horvath (2006) that increases the
dissimilarity between two individuals to one if they don't belong to the
same terminal node. The second approach of Fouodo et al. (2021) takes the tree
depth into account and estimate the dissimilarity between two individuals
basing on the length of the minimal path between the terminal nodes they
belong to.
}
\examples{
\dontrun{
set.seed(1234)
index <- sample(x = 1:150, size = 15, replace = FALSE)
  iris_dissimilarities <- pranger(
                             data = iris[ , -5],
                             strategy = "boostrepl")
}
}
\author{
Cesaire J. K. Fouodo
}
