% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pranger.R
\name{pranger}
\alias{pranger}
\title{Title}
\usage{
pranger(data, strategy, verbose, ...)
}
\arguments{
\item{data}{[data.frame(1)] A \code{data.frame} of original dataset}

\item{strategy}{[character(1)] Name of the resampling strategy to be used. Most be element of
"boostrepl", "boostwithoutrepl", "boostbayes",
"unif", "normal" or "binomial"}

\item{verbose}{[boolean] If TRUE, verbose}

\item{...}{further parameters to be passed to \code{ranger}}
}
\value{
[matrix(1)] Matrix of dissimilarities
}
\description{
Title
}
\examples{
\dontrun{
set.seed(1234)
index <- sample(x = 1:150, size = 15, replace = FALSE)
  iris_dissimilarities <- pranger(
                             data = iris[ , -5],
                             strategy = "boostrepl")
}
}
\author{
Cesaire J. K. Fouodo
}
