% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predicted_forest_distance.R
\name{predicted_forest_distance}
\alias{predicted_forest_distance}
\title{predicted_forest_distance Compute dissimilarity matrix over the forest. Node
that this function will parallelize over the available CPUs nodes.}
\usage{
predicted_forest_distance(predictions, forest)
}
\arguments{
\item{predictions}{[ranger(1)] An ranger predicted object of terminal nodes}

\item{forest}{[ranger(1)] A ranger forest}
}
\value{
[matrix(1)] Matrix of dissimilarities over the forest
}
\description{
predicted_forest_distance Compute dissimilarity matrix over the forest. Node
that this function will parallelize over the available CPUs nodes.
}
\examples{
\dontrun{
set.seed(1234)
train.idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris.train <- iris[train.idx, ]
iris.test <- iris[-train.idx, ]
rg.iris <- ranger(Species ~ ., data = iris.train)
pred.iris <- predict(rg.iris, data = iris.test, type = "terminalNodes")
 iris_dist <- predicted_forest_distance(
   forest = rg.iris,
   predictions = pred.iris)
}
}
\author{
Cesaire J. K. Fouodo
}
