% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_predicted_forest_one_tree.R
\name{predicted_forest_distance_parall}
\alias{predicted_forest_distance_parall}
\title{predicted_forest_distance_parall computes dissimilarities for one tree in the forest}
\usage{
predicted_forest_distance_parall(index_tree, predictions, forest)
}
\arguments{
\item{index_tree}{[integer(1)] Index of concerned tree}

\item{predictions}{[ranger(1)] An ranger predicted object of terminal nodes}

\item{forest}{[ranger(1)] A ranger forest}
}
\value{
[matrix(1)] Matrix of dissimilarities for the concerned tree
}
\description{
predicted_forest_distance_parall computes dissimilarities for one tree in the forest
}
\author{
Cesaire J. K. Fouodo
}
