% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predicted_tree_distance.R
\name{predicted_tree_distance}
\alias{predicted_tree_distance}
\title{predicted_tree_distance uses predicted terminal nodes of a tree to compute
the dissimilarity matrix}
\usage{
predicted_tree_distance(tree_index, forest, predictions)
}
\arguments{
\item{tree_index}{[integer(1)] A ranger tree}

\item{forest}{[ranger(1)] A ranger forest}

\item{predictions}{[ranger(1)] Predictions object from class ranger}
}
\value{
[matrix(1)] A matrix of dissimilarities
}
\description{
predicted_tree_distance uses predicted terminal nodes of a tree to compute
the dissimilarity matrix
}
\examples{
set.seed(1234)
train.idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris.train <- iris[train.idx, ]
iris.test <- iris[-train.idx, ]
rg.iris <- ranger(Species ~ ., data = iris.train)
pred.iris <- predict(rg.iris, data = iris.test, type = "terminalNodes")
 iris_dist <- predicted_tree_distance(
   tree_index = 1,
   forest = rg.iris,
   predictions = pred.iris)
}
\author{
Cesaire J. K. Fouodo
}
