% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resampling}
\alias{resampling}
\title{Pranger}
\usage{
resampling(data, strategy, nb_bootst, aggregation)
}
\arguments{
\item{data}{[\code{data.frame}] A \code{data.frame} of original dataset}

\item{strategy}{[\code{character}] Name of strategy to be used. Most be element of
"boostrepl", "boostwithoutrepl", "boostbayes",
"unif", "normal", "binomial" or "boostaggr"}

\item{nb_bootst}{[\code{integer}] Number of repetitions required to aggregate the
bootstrap samples. Set to ceiling(sqrt(n)) if not
provided, with n the number of observations}

\item{aggregation}{[\code{function}] Function that is called for aggregation. Only
usable in the case of "boostaggr"}
}
\value{
[\code{data.frame}] A \code{data.frame} of a synthetic two-classes classification
                        problem
}
\description{
Synthesize a two-classes classification problem given a resampling
           strategy.
}
\examples{
    set.seed(1234)
   data_synt <- resampling(
                    data = data.frame(x = rnorm(10),
                                      y = rnorm(10)),
                    strategy = "boostrepl"
                  )
}
\author{
Cesaire J. K. Fouodo
}
